+++ pybv/io.py
@@ -11,7 +11,6 @@
 #
 # License: BSD-3-Clause
 
-import codecs
 import datetime
 import os
 import shutil
@@ -35,7 +34,10 @@ SUPPORTED_VOLTAGE_SCALINGS = {
 }
 
 
-def write_brainvision(*, data, sfreq, ch_names, fname_base, folder_out,
+def write_brainvision(*, data, sfreq, ch_names,
+                      ref_ch_names=None,
+                      fname_base,
+                      folder_out,
                       overwrite=False,
                       events=None,
                       resolution=0.1,
@@ -53,7 +55,17 @@ def write_brainvision(*, data, sfreq, ch_names, fname_base, folder_out,
     sfreq : int | float
         The sampling frequency of the data.
     ch_names : list of {str | int}, len (n_channels)
-        The name of each channel.
+        The names of the channels.
+    ref_ch_names : str | list of str, len (n_channels) | None
+        The name of the channel used as a reference during the recording. If
+        references differed between channels, you may supply a list of
+        reference channel names corresponding to each channel in ``ch_names``.
+        If ``None`` (default), assume that all channels are referenced to a
+        common channel that is not further specified (BrainVision default).
+
+        .. note:: The reference channel name specified here does not need to
+                  appear in ``ch_names``. It is permissible to specify a
+                  reference channel that is not present in ``data``.
     fname_base : str
         The base name for the output files. Three files will be created
         (.vhdr, .vmrk, .eeg) and all will share this base name.
@@ -161,6 +173,38 @@ def write_brainvision(*, data, sfreq, ch_names, fname_base, folder_out,
     if len(set(ch_names)) != nchan:
         raise ValueError("Channel names must be unique, found duplicate name.")
 
+    # Ensure we have a list of strings as reference channel names
+    if ref_ch_names is None:
+        ref_ch_names = [''] * nchan  # common but unspecified reference
+    elif isinstance(ref_ch_names, str):
+        ref_ch_names = [ref_ch_names] * nchan
+    else:
+        if "" in ref_ch_names:
+            msg = (f"ref_ch_names contains an empty string: {ref_ch_names}\n"
+                   f"Empty strings are reserved values and not permitted "
+                   f"as reference channel names.")
+            raise ValueError(msg)
+        ref_ch_names = [str(ref_ch_name) for ref_ch_name in ref_ch_names]
+
+    if len(ref_ch_names) != nchan:
+        raise ValueError(
+            f'The number of reference channel names ({len(ref_ch_names)})'
+            f'must match the number of channels in your data ({nchan})'
+        )
+
+    # ensure ref chs that are in data are zero
+    for ref_ch_name in list(set(ref_ch_names) & set(ch_names)):
+        if not np.allclose(data[ch_names.index(ref_ch_name), :], 0):
+            raise ValueError(
+                f"The provided data for the reference channel "
+                f"{ref_ch_name} does not appear to be zero across "
+                f"all time points. This indicates that this channel "
+                f"either did not serve as a reference during the recording, "
+                f"or the data has been altered since. Please either pick a "
+                f"different reference channel, or omit the "
+                f"ref_ch_name parameter."
+            )
+
     if not isinstance(sfreq, (int, float)):
         raise ValueError("sfreq must be one of (float | int)")
     sfreq = float(sfreq)
@@ -230,9 +274,11 @@ def write_brainvision(*, data, sfreq, ch_names, fname_base, folder_out,
         _write_bveeg_file(eeg_fname, data, orientation='multiplexed',
                           format=fmt, resolution=resolution, units=units)
         _write_vmrk_file(vmrk_fname, eeg_fname, events, meas_date)
-        _write_vhdr_file(vhdr_fname, vmrk_fname, eeg_fname, data, sfreq,
-                         ch_names, orientation='multiplexed', format=fmt,
-                         resolution=resolution, units=units)
+        _write_vhdr_file(vhdr_fname=vhdr_fname, vmrk_fname=vmrk_fname,
+                         eeg_fname=eeg_fname, data=data, sfreq=sfreq,
+                         ch_names=ch_names, ref_ch_names=ref_ch_names,
+                         orientation='multiplexed',
+                         format=fmt, resolution=resolution, units=units)
     except ValueError:
         if folder_out_created:
             # if this is a new folder, remove everything
@@ -266,7 +312,7 @@ def _chk_multiplexed(orientation):
 
 def _write_vmrk_file(vmrk_fname, eeg_fname, events, meas_date):
     """Write BrainvVision marker file."""
-    with codecs.open(vmrk_fname, 'w', encoding='utf-8') as fout:
+    with open(vmrk_fname, 'w', encoding='utf-8') as fout:
         print('Brain Vision Data Exchange Marker File, Version 1.0', file=fout)
         print(f';Exported using pybv {__version__}', file=fout)
         print('', file=fout)
@@ -348,14 +394,15 @@ def _scale_data_to_unit(data, units):
     return data * scales
 
 
-def _write_vhdr_file(vhdr_fname, vmrk_fname, eeg_fname, data, sfreq, ch_names,
-                     orientation, format, resolution, units):
+def _write_vhdr_file(*, vhdr_fname, vmrk_fname, eeg_fname, data, sfreq,
+                     ch_names, ref_ch_names, orientation, format, resolution,
+                     units):
     """Write BrainvVision header file."""
     bvfmt, _ = _chk_fmt(format)
 
     multiplexed = _chk_multiplexed(orientation)
 
-    with codecs.open(vhdr_fname, 'w', encoding='utf-8') as fout:
+    with open(vhdr_fname, 'w', encoding='utf-8') as fout:
         print('Brain Vision Data Exchange Header File Version 1.0', file=fout)
         print(f'; Written using pybv {__version__}', file=fout)
         print('', file=fout)
@@ -392,10 +439,14 @@ def _write_vhdr_file(vhdr_fname, vmrk_fname, eeg_fname, data, sfreq, ch_names,
         resolutions = resolution * np.ones((nchan,))
 
         for i in range(nchan):
+            # take care of commas in the channel names
             _ch_name = ch_names[i].replace(',', r'\1')
+            _ref_ch_name = ref_ch_names[i].replace(',', r'\1')
+
             resolution = np.format_float_positional(resolutions[i], trim="-")
             unit = units[i]
-            print(f'Ch{i + 1}={_ch_name},,{resolution},{unit}', file=fout)
+            print(f'Ch{i + 1}={_ch_name},{_ref_ch_name},{resolution},{unit}',
+                  file=fout)
 
         print('', file=fout)
         print('[Comment]', file=fout)